packages/ml-training/pyproject.toml
[tool.poetry]
name = "darwin-education-ml-training"
version = "0.1.0"
description = "ML training pipelines for Darwin Education."
authors = ["Darwin Education <dev@darwin.education>"]
readme = "README.md"
packages = [{ include = "darwin_ml", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
lightgbm = "^4.3.0"
onnx = "^1.16.0"
onnxruntime = "^1.17.3"
onnxmltools = "^1.12.0"
supabase = "^2.4.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
ruff = "^0.4.7"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
